[MOD] Del use PWA in favour of layout
